requirements.txt
streamlit
PyPDF2
pdfplumber
google.generativeai
python-dotenv
pandas
extract_msg
python-docx
pymongo